Move farmer drop & thread utils to networking utils
