+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/cluster/controller.rs
@@ -27,9 +27,9 @@ use subspace_farmer::node_client::NodeClient;
 use subspace_farmer::node_client::caching_proxy_node_client::CachingProxyNodeClient;
 use subspace_farmer::node_client::rpc_node_client::RpcNodeClient;
 use subspace_farmer::single_disk_farm::identity::Identity;
-use subspace_farmer::utils::{AsyncJoinOnDrop, run_future_in_dedicated_thread};
 use subspace_kzg::Kzg;
 use subspace_networking::utils::piece_provider::PieceProvider;
+use subspace_networking::utils::{AsyncJoinOnDrop, run_future_in_dedicated_thread};
 use tracing::{Instrument, info, info_span};
 
 /// Get piece retry attempts number.
