+++ crates/subspace-farmer/src/cluster/farmer.rs
@@ -15,7 +15,6 @@ use crate::farm::{
     Farm, FarmError, FarmId, FarmingNotification, HandlerFn, HandlerId, PieceReader,
     PlottedSectors, SectorUpdate,
 };
-use crate::utils::AsyncJoinOnDrop;
 use anyhow::anyhow;
 use async_trait::async_trait;
 use derive_more::{Display, From};
@@ -31,6 +30,7 @@ use std::time::{Duration, Instant};
 use subspace_core_primitives::pieces::{Piece, PieceOffset};
 use subspace_core_primitives::sectors::SectorIndex;
 use subspace_farmer_components::plotting::PlottedSector;
+use subspace_networking::utils::AsyncJoinOnDrop;
 use subspace_rpc_primitives::SolutionResponse;
 use tokio::time::MissedTickBehavior;
 use tracing::{Instrument, debug, error, info_span, trace, warn};
