+++ crates/subspace-farmer/src/disk_piece_cache.rs
@@ -8,7 +8,6 @@ use crate::disk_piece_cache::metrics::DiskPieceCacheMetrics;
 use crate::farm;
 use crate::farm::{FarmError, PieceCacheId, PieceCacheOffset};
 use crate::single_disk_farm::direct_io_file::{DISK_SECTOR_SIZE, DirectIoFile};
-use crate::utils::AsyncJoinOnDrop;
 use async_trait::async_trait;
 use bytes::BytesMut;
 use futures::channel::mpsc;
@@ -24,6 +23,7 @@ use std::{fs, io};
 use subspace_core_primitives::hashes::{Blake3Hash, blake3_hash_list};
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
 use subspace_farmer_components::file_ext::FileExt;
+use subspace_networking::utils::AsyncJoinOnDrop;
 use thiserror::Error;
 use tokio::runtime::Handle;
 use tokio::task;
