+++ crates/subspace-farmer/src/plotter/cpu.rs
@@ -5,7 +5,6 @@ pub mod metrics;
 use crate::plotter::cpu::metrics::CpuPlotterMetrics;
 use crate::plotter::{Plotter, SectorPlottingProgress};
 use crate::thread_pool_manager::PlottingThreadPoolManager;
-use crate::utils::AsyncJoinOnDrop;
 use async_lock::{Mutex as AsyncMutex, Semaphore, SemaphoreGuardArc};
 use async_trait::async_trait;
 use bytes::Bytes;
@@ -35,6 +34,7 @@ use subspace_farmer_components::plotting::{
     encode_sector, write_sector,
 };
 use subspace_kzg::Kzg;
+use subspace_networking::utils::AsyncJoinOnDrop;
 use subspace_proof_of_space::Table;
 use tokio::task::yield_now;
 use tracing::{Instrument, warn};
